refactor(download): extract constants and move platform types to a types file
Address review feedback and the types-location CI gate:

- api: move the desktop-release constants out of the service into app/constants
  (GitHub repo, tag prefix, page size, timeout in constants/desktop.py; cache key
  + TTL in constants/cache.py alongside the other desktop bridge entries).
- web: move the platform/desktop type and interface exports out of
  usePlatformDetection.ts into a colocated usePlatformDetection.types.ts and
  re-export them, so the hook no longer exports >3 types (CI: "Types live in
  dedicated *.types.ts files").

Entire-Checkpoint: 3f934f7ecbba
