+++ apps/web/src/hooks/ui/usePlatformDetection.ts
@@ -3,37 +3,21 @@
 import { useEffect, useMemo, useState } from "react";
 import { desktopApi } from "@/features/download/api/desktopApi";
 import type { DesktopRelease } from "@/features/download/types";
+import type {
+  DesktopArch,
+  DesktopOS,
+  DesktopVariant,
+  Platform,
+  PlatformInfo,
+} from "./usePlatformDetection.types";
 
-export type Platform =
-  | "mac-arm"
-  | "mac-intel"
-  | "windows"
-  | "linux"
-  | "ios"
-  | "android"
-  | "unknown";
-
-export type DesktopOS = "mac" | "windows" | "linux";
-export type DesktopArch = "x64" | "arm64";
-
-export interface PlatformInfo {
-  platform: Platform;
-  displayName: string;
-  shortName: string;
-  iconPath: string;
-  downloadUrl: string | null;
-  isDesktop: boolean;
-  isMobile: boolean;
-}
-
-export interface DesktopVariant {
-  os: DesktopOS;
-  arch: DesktopArch;
-  label: string;
-  description: string;
-  /** Direct asset URL, or null when this binary isn't published in the release. */
-  downloadUrl: string | null;
-}
+export type {
+  DesktopArch,
+  DesktopOS,
+  DesktopVariant,
+  Platform,
+  PlatformInfo,
+} from "./usePlatformDetection.types";
 
 const GITHUB_RELEASES_BASE =
   "https://github.com/theexperiencecompany/gaia/releases";
